requirements.txt
discord.py
python-dotenv
pyyaml
requests
slack_bolt
